update export python code with states
