+++ .gitignore
@@ -163,3 +163,6 @@ cython_debug/
 
 # vscode test
 .vscode
+
+# default results folder
+results
